Generalize TCP over IP endpoint implementation
Resolves #672 (Generalize TCP over IP endpoint implementation).
